+++ Sandswept/Items/Lunars/HallowedIchor.cs
@@ -42,7 +42,7 @@ namespace Sandswept.Items.Lunars
         [ConfigField("Chest Reopen Difficulty Coefficient Multiplier Add", "Just check the Formula Example..", 0.1f)]
         public static float chestReopenDifficultyCoefficientMultiplierAdd;
 
-        [ConfigField("Per Player Divisor Add", "Just check the Formula Example...", 0.25f)]
+        [ConfigField("Per Player Divisor Add", "Just check the Formula Example...", 0.175f)]
         public static float perPlayerDivisorAdd;
 
         [ConfigField("Chest Reopen Difficulty Coefficient Flat Add Scalar", "Just check the Formula Example....", 0.06f)]
@@ -76,6 +76,7 @@ namespace Sandswept.Items.Lunars
         public static Color32 cachedTimerColor = Color.white;
 
         public static bool anyoneHadHallowedIchor = false;
+        public static bool anyoneHadHallowedIchorEver = false;
 
         public static int globalReopenCount = 0;
 
@@ -193,7 +194,7 @@ namespace Sandswept.Items.Lunars
         // the code may be unoptimized buuut it doesn't really take up performance
         private void PerformJitterUI(On.RoR2.UI.RunTimerUIController.orig_Update orig, RoR2.UI.RunTimerUIController self)
         {
-            if (anyoneHadHallowedIchor)
+            if (anyoneHadHallowedIchorEver)
             {
                 double time = 0f;
 
@@ -251,6 +252,7 @@ namespace Sandswept.Items.Lunars
             var runDifficultyDef = DifficultyCatalog.GetDifficultyDef(run.selectedDifficulty);
             runDifficultyDef.scalingValue = cachedDifficultyDefScalingValue;
             anyoneHadHallowedIchor = false;
+            anyoneHadHallowedIchorEver = false;
             globalReopenCount = 0;
         }
 
@@ -269,6 +271,7 @@ namespace Sandswept.Items.Lunars
             if (itemCount > 0)
             {
                 anyoneHadHallowedIchor = true;
+                anyoneHadHallowedIchorEver = true;
             }
             else
             {
@@ -415,7 +418,49 @@ namespace Sandswept.Items.Lunars
 
         public override ItemDisplayRuleDict CreateItemDisplayRules()
         {
-            return new ItemDisplayRuleDict();
+            var itemDisplay = SetUpIDRS();
+
+            ItemDisplayRuleDict i = new();
+
+            #region Sandswept Survivors
+
+            i.Add("RangerBody",
+
+                new ItemDisplayRule()
+                {
+                    ruleType = ItemDisplayRuleType.ParentedPrefab,
+                    childName = "Chest",
+                    localPos = new Vector3(-0.00387F, 0.11857F, 0.01629F),
+                    localAngles = new Vector3(84.61184F, 220.3867F, 47.41245F),
+                    localScale = new Vector3(0.14531F, 0.14659F, 0.14531F),
+
+                    followerPrefab = itemDisplay,
+                    limbMask = LimbFlags.None,
+                    followerPrefabAddress = new AssetReferenceGameObject("")
+                }
+
+            );
+
+            i.Add("ElectricianBody",
+
+                new ItemDisplayRule()
+                {
+                    ruleType = ItemDisplayRuleType.ParentedPrefab,
+                    childName = "Head",
+                    localPos = new Vector3(-0.01041F, 0.08162F, -0.00924F),
+                    localAngles = new Vector3(85.0407F, 197.8464F, 22.78797F),
+                    localScale = new Vector3(0.12683F, 0.11843F, 0.11843F),
+
+                    followerPrefab = itemDisplay,
+                    limbMask = LimbFlags.None,
+                    followerPrefabAddress = new AssetReferenceGameObject("")
+                }
+
+            );
+
+            #endregion
+
+            return i;
         }
     }