+++ Sandswept/Mechanics/NewtReflection.cs
@@ -1,3 +1,4 @@
+/*
 using System;
 using Sandswept.Artifacts;
 using Sandswept.Buffs;
@@ -184,3 +185,4 @@ namespace Sandswept.Mechanics {
         }
     }
 }
+*/